把全局的 flatf 重命名为 flat，并将参数类型从 float 改为 CGFloat，从而修复 UIImage 绘制图片时由于精度不足导致可能尺寸不正确的 bug

+++ QMUIKit/UIComponents/ImagePickerLibrary/QMUIImagePickerViewController.m
@@ -257,7 +257,7 @@ static QMUIImagePickerViewController *imagePickerViewControllerAppearance;
 - (CGSize)referenceImageSize {
     CGFloat collectionViewWidth = CGRectGetWidth(self.collectionView.bounds);
     CGFloat collectionViewContentSpacing = collectionViewWidth - UIEdgeInsetsGetHorizontalValue(self.collectionView.contentInset);
-    NSInteger columnCount = floorf(collectionViewContentSpacing / self.minimumImageWidth);
+    NSInteger columnCount = floor(collectionViewContentSpacing / self.minimumImageWidth);
     CGFloat referenceImageWidth = self.minimumImageWidth;
     BOOL isSpacingEnoughWhenDisplayInMinImageSize = UIEdgeInsetsGetHorizontalValue(self.collectionViewLayout.sectionInset) + (self.minimumImageWidth + self.collectionViewLayout.minimumInteritemSpacing) * columnCount - self.collectionViewLayout.minimumInteritemSpacing <= collectionViewContentSpacing;
     if (!isSpacingEnoughWhenDisplayInMinImageSize) {
