+++ QMUIKit/UIKitExtensions/UICollectionView+QMUI.m
@@ -307,7 +307,7 @@
         }
         fitSize = [cell sizeThatFits:CGSizeMake(itemWidth, CGFLOAT_MAX)];
     }
-    return ceilf(fitSize.height);
+    return ceil(fitSize.height);
 }
 
 // 通过indexPath缓存高度
